:bath: couldbuild: Remove unnecessary grep option
Remove the `-s` grep option in the `cover` cloudbuild stage as it does
nothing useful. It only makes sense when grep is reading from a file,
not stdin. Originally I meant `-q` to suppress grep's output, but on
consideration I think I like it that grep outputs 100.0% coverage on
success. So no flags for grep are needed.

Issue: #14