:sparkles: cloudbuild: Enable race detector on tests
Go has a race detector enabled by running `go` with `-race`. This was
disabled for Google cloudbuild because it didn't work because the
cloud-builder used was alpine-based, and the race detector does not work
on alpine (golang go issues 14481). The race detector needs glibc but
alpine uses musl.

Switch to a debian-based cloud-builder image instead. Debian has glibc
so the race detector works.

Issues: #14