:sparkles: travis: Unify travis and cloudbuild steps
The travis config and the cloudbuild config had different commands for
the coverage checking step. They both work fine, but it makes sense to
use the same command in both configs so as to remove any question of why
they might be different.

This change slightly alters the command in that it will output the
`total: ... 100.0%` line on success, whereas previously it was silent on
success. This was done for cloudbuild so we're doing the same here.

This also changes the name of the coverage output file to `coverage.out`
from `coverage.txt`, again keeping it the same as the Google cloud build
config. No reason to maintain that difference in filenames.

Issue: #14