Cloudbuild cleanup (#15)
* :bath: couldbuild: Remove unnecessary grep option

Remove the `-s` grep option in the `cover` cloudbuild stage as it does
nothing useful. It only makes sense when grep is reading from a file,
not stdin. Originally I meant `-q` to suppress grep's output, but on
consideration I think I like it that grep outputs 100.0% coverage on
success. So no flags for grep are needed.

* :sparkles: cloudbuild: Enable race detector on tests

Go has a race detector enabled by running `go` with `-race`. This was
disabled for Google cloudbuild because it didn't work because the
cloud-builder used was alpine-based, and the race detector does not work
on alpine (golang go issues 14481). The race detector needs glibc but
alpine uses musl.

Switch to a debian-based cloud-builder image instead. Debian has glibc
so the race detector works.

* :sparkles: travis: Unify travis and cloudbuild steps

The travis config and the cloudbuild config had different commands for
the coverage checking step. They both work fine, but it makes sense to
use the same command in both configs so as to remove any question of why
they might be different.

This change slightly alters the command in that it will output the
`total: ... 100.0%` line on success, whereas previously it was silent on
success. This was done for cloudbuild so we're doing the same here.

This also changes the name of the coverage output file to `coverage.out`
from `coverage.txt`, again keeping it the same as the Google cloud build
config. No reason to maintain that difference in filenames.

Issues: #14