Extends Storer interface for all methods to also return an error
In order to implement the Storer interface, all methods must now return
and error.

The value property in Puppy has had its type changed to an int to
represent a numerical cent figure so that it may be properly throw an
error if the value supplied is less than 0.

Lab 07 (#589)