travis: Bump Go version to 1.13
Go 1.13 has been released. Normally I would not bump the version on the
last day of the course, but it looks like the golanci lint service is
now running with Go 1.13, and might be causing spurious linting errors.

So bump the version of Go. A subsequent commit will bump the version of
golanci-lint, so we can see if this fixes the problems we're seeing.