lint: Bump golangci-lint version to v1.19.1
The golangci lint service is running Go 1.13 now, so we've bumped the
version of Go. We are seeing weird errors with the linter service that
may be related to v1.17.1 of the linter on Go 1.13. So bump the linter
version to the latest to see if that resolves the problem.

Errors are:

    /usr/local/go/src/fmt/scan.go:1200:16: DecodeRuneInString not declared by package utf8
    /usr/local/go/src/encoding/json/fold.go:83:20: DecodeRune not declared by package utf8
    /usr/local/go/src/strings/strings.go:980:20: DecodeRuneInString not declared by package utf8
    /go/pkg/mod/github.com/stretchr/testify@v1.4.0/assert/assertions.go:174:18: DecodeRuneInString not declared by package utf8

The utf8 package definitely has those functions. It is not clear why
we're getting these errors, or even why lint is running over the Go
standard library.

Disable the funlen, godox and whitespace linters. They are causing
unnecessary errors. I would like to turn on the whitespace linter, but
there are too many existing errors.

Increase the deadline for `golangci-lint run` on travis as it looks like
with the new linters on this code base, we're hitting the 1 minute
default. Increase it to 2 minutes, which should give us plenty of
headroom.